Enhances cache functionality
When using "php" as extension property the cachobject will be stored as
returning array in a php file. This array contains all of the objects
properties. The objects content is base64 encoded to prevent
syntaxerrors.

Other extensions then "php" are the extension of the cachefile to
create. This feature for example is used for combined css and js files.

Signed-off-by: Michael "Tekkla" Zorn <tekkla@tekkla.de>