Adds css and js combining and caching
On getCss() and  getScript() now all local css/js files and inline
styles/scripts are combined and stored in filecache. The combining
reduces the number requests and the amount of data to be transfered
while filecaching reduces needed cpu power. Default TTL for js and css
files is 3600s and can be changed via Core config.

Signed-off-by: Michael "Tekkla" Zorn <tekkla@tekkla.de>