Simplifies code
Removes the Core\lib\Amvc\Creator::create() method in favor of
getAppinstance() which did the same job. All references in other classes
have been updated.

Removes also the clone feature from getAppInstance() which was a relict
from a time where this framework was planned to be integrated into
another software which needed this.

Adds more filename and classname checks in getAppInstance() to prevent
errors when calling an app, controller or action that does not exist.
There is still sometimes an error where getAppInstance() is called with
an empty name. Until this error has been found some debug code stays in
getAppInstance() method.

Signed-off-by: Michael "Tekkla" Zorn <tekkla@tekkla.de>