Massive change of route mapping
The way how routes get mapped has been changed fundamentally. Now,
routes have to be defined as key based array. This key will be used as
name and as default value for controller and action if no target has
been set in routeconfig.

The route mapping logic in App::initRoutes() was moved into
Router::mapAppRoutes() method to make it accessible from outside apps.
This new method expects the name of the app the routes are from and a
key based array with routes as descripbed above.

When no "route" element is set in a routes config, the routes name will
be exploded (.) and the first element used to create a route like
"/firste_element".

When no "target" element is set in a routes config, the exploded first
part of a routes name will be used as controllername and a possible
second part as actionname. When there is no exploded second part, the
fist part will be used as actioname too. A routename like "fist.second"
will set "first" as controllername and "second" as actioname as routes
targetconfig.

New method Router::getRoutes() to return all registered routes. This is
mainly for debugging purpose.

Signed-off-by: Michael Tekkla Zorn <tekkla@tekkla.de>